Relicense docs and metadata to CC-BY
I need to explain myself here. For a moment, I thought CC0 would be the best option, as we are dealing with mostly non-copyrightable content (metadata) and it would be most permissive. However, later I realized that external contributors may want at least an attribution of their input to the docs (which definitely ARE copyrightable).

I realize that the docs and metadata were technically available for a few hours under CC0 and thus are licensed as CC0 forever. I'm completely fine with that :) However, newer versions of the docs and the metadata will be CC BY 4.0 only.